SigningService: Split tests into several files
